infrastructure/requirements.txt
aws-cdk-lib==2.120.0
constructs>=10.0.0,<11.0.0
boto3==1.34.34
opensearch-py==2.4.2
requests-aws4auth==1.2.3
